requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
autopep8==1.5.7
beautifulsoup4==4.10.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.6
django-allauth==0.45.0
django-cors-headers==3.8.0
django-crispy-forms==1.12.0
graphene==2.1.9
graphene-django==2.14.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.2
numpy==1.21.4
oauthlib==3.1.1
pandas==1.3.4
promise==2.3
psycopg2-binary==2.9.1
pycodestyle==2.8.0
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.6
